Change safe_copy_str's nSizeOfBuffer argument from int to size_t.
git-svn-id: https://tinymux.googlecode.com/svn/branches/dev_jake@751 d1b986fa-651c-0410-a323-35a8662cf44d

+++ mux/src/stringutil.cpp
@@ -1938,7 +1938,7 @@ char *BufferCloneLen(const char *pBuffer, unsigned int nBuffer)
  * safe_copy_str, safe_copy_chr - Copy buffers, watching for overflows.
  */
 
-void safe_copy_str(const char *src, char *buff, char **bufp, int nSizeOfBuffer)
+void safe_copy_str(const char *src, char *buff, char **bufp, size_t nSizeOfBuffer)
 {
     if (src == NULL) return;
